153 Overcooked Integration (#168)
* space serialization

* web app

* linters

* linters

* linters

* gitignore

* model_registry_v2 and smoke tests

* remove smoke tests

* remove changes from Smoke test PR

* remove python-version

* self-review

* self-review

* self-review

* linters

* linters

* use pysdk model_registry_v2

* remove old model registry

* fix config

* fix config

* self review

* adapt new actors

* lintersd

* fix pong impl

* soft-actor critic test config

* base integration

* remove overcooked_ai

* requirements fork

* random experiment

* overcooked turn based controls

* web app build

* fix UI integration

* fix random experiment

* linters

* linters

* add overcooked requirements to gitlab

* Update overcooked_ai.md

* address PR comments

* fix licenses

---------

+++ cogment_verse/specs/spaces_serialization.py
@@ -37,9 +37,9 @@ def serialize_gym_space(gym_space, serilization_format=SerializationFormat.STRUC
         if isinstance(gym_space.n, np.ndarray):
             size = gym_space.n
         elif isinstance(gym_space.n, int):
-            size = np.array([gym_space.n], dtype=np.dtype("int8"))
+            size = np.array([gym_space.n], dtype=np.dtype("int32"))
         else:
-            size = np.array(gym_space.n, dtype=np.dtype("int8"))
+            size = np.array(gym_space.n, dtype=np.dtype("int32"))
         return Space(multi_binary=MultiBinary(n=serialize_ndarray(size, serilization_format=serilization_format)))
 
     if isinstance(gym_space, (gym.spaces.MultiDiscrete, gymnasium.spaces.MultiDiscrete)):
